requirements.txt
dotmap==1.3.26
kornia==0.6.3
numpy==1.21.2
Pillow==9.0.1
pytorch_lightning==0.9.0
torch==1.10.1
torchmetrics==0.7.2
torchvision==0.11.2
